requirements.txt
requests
pytest
pytest-html
pytest-mock
pyyaml
allure-pytest
wiremock
jsonschema
uritemplate